+++ src/NeuralNet/Initializers/Constant.php
@@ -3,7 +3,11 @@
 namespace Rubix\ML\NeuralNet\Initializers;
 
 use Tensor\Matrix;
+use Tensor\ColumnVector;
+use Rubix\ML\NeuralNet\Parameter;
 use Rubix\ML\Exceptions\InvalidArgumentException;
+
+use function count;
 
 /**
  * Constant
@@ -38,17 +42,35 @@ class Constant implements Initializer
     }
 
     /**
-     * Initialize a weight matrix W in the dimensions fan in x fan out.
+     * Initialize a parameter tensor given its target shape.
      *
      * @internal
      *
-     * @param int<0,max> $fanIn
-     * @param int<0,max> $fanOut
-     * @return Matrix
+     * @param int[] $shape
+     * @return Parameter
      */
-    public function initialize(int $fanIn, int $fanOut) : Matrix
+    public function initialize(array $shape) : Parameter
     {
-        return Matrix::fill($this->value, $fanOut, $fanIn);
+        switch (count($shape)) {
+            case 1:
+                $n = $shape[0];
+
+                $tensor = ColumnVector::fill($this->value, $n);
+
+                break;
+
+            case 2:
+                [$m, $n] = $shape;
+
+                $tensor = Matrix::fill($this->value, $m, $n);
+
+                break;
+
+            default:
+                throw new InvalidArgumentException('Invalid shape for Constant initializer.');
+        }
+
+        return new Parameter($tensor);
     }
 
     /**