+++ tests/Classifiers/LogisticRegressionTest.php
@@ -106,6 +106,14 @@ class LogisticRegressionTest extends TestCase
     }
 
     #[Test]
+    public function negativeL1Penalty() : void
+    {
+        $this->expectException(InvalidArgumentException::class);
+
+        new LogisticRegression(l1Penalty: -1.0);
+    }
+
+    #[Test]
     public function type() : void
     {
         $this->assertEquals(EstimatorType::classifier(), $this->estimator->type());
@@ -127,6 +135,7 @@ class LogisticRegressionTest extends TestCase
         $expected = [
             'batch size' => 100,
             'optimizer' => new Adam(new Constant(0.01)),
+            'l1 penalty' => 1e-4,
             'l2 penalty' => 1e-4,
             'epochs' => 300,
             'min change' => 1e-4,